Rate-limit ros spin loops
Previously, two loops each saturate a cpu core to process messages with
as low latency as possible. While this design works for real-time
performance, our deployment scenarios are power-constrained and have
other nodes running.

These changes reduce the average cpu utilization of the node (after
initialization) from 200% to 4%. HuskySystem reports that the filter
is no longer running in real-time, with an average queue size of 2.